Finishing up getting rid of all the compilation warnings with clang.

+++ src/narf/client/world.cpp
@@ -129,7 +129,6 @@ void narf::client::World::render(narf::gl::Texture *tiles_tex, const narf::Camer
 	 *   F  |  G  |  H
 	 */
 
-	// TODO: Would be better if these were uint32_t, so maybe do the assert differently
 	int32_t cx_min = cx - std::min(renderDistance, (int32_t)chunks_x_ - 1);
 	int32_t cx_max = cx + std::min(renderDistance, (int32_t)chunks_x_ - 1) + 1;
 
@@ -139,17 +138,17 @@ void narf::client::World::render(narf::gl::Texture *tiles_tex, const narf::Camer
 	assert(cx_max >= 0);
 	assert(cy_max >= 0);
 
-	uint32_t mid_cx_min = std::max(0, cx_min);
-	uint32_t mid_cy_min = std::max(0, cy_min);
+	uint32_t mid_cx_min = static_cast<uint32_t>(std::max(0, cx_min));
+	uint32_t mid_cy_min = static_cast<uint32_t>(std::max(0, cy_min));
 
-	uint32_t mid_cx_max = std::min((int32_t)chunks_x_, cx_max);
-	uint32_t mid_cy_max = std::min((int32_t)chunks_y_, cy_max);
+	uint32_t mid_cx_max = static_cast<uint32_t>(std::min((int32_t)chunks_x_, cx_max));
+	uint32_t mid_cy_max = static_cast<uint32_t>(std::min((int32_t)chunks_y_, cy_max));
 
 	renderSlice(tiles_tex, mid_cx_min, mid_cx_max, mid_cy_min, mid_cy_max, stateBlend);
 
 	if ((uint32_t)cx_max > chunks_x_) {
 		uint32_t a_cx_min = 0;
-		uint32_t a_cx_max = cx_max - chunks_x_;
+		uint32_t a_cx_max = static_cast<uint32_t>(cx_max) - chunks_x_;
 		glPushMatrix();
 		glTranslatef((float)size_x_, 0.0f, 0.0f);
 		renderSlice(tiles_tex, a_cx_min, a_cx_max, mid_cy_min, mid_cy_max, stateBlend);
@@ -157,7 +156,7 @@ void narf::client::World::render(narf::gl::Texture *tiles_tex, const narf::Camer
 
 		if ((uint32_t)cy_max > chunks_y_) {
 			uint32_t b_cy_min = 0;
-			uint32_t b_cy_max = cy_max - chunks_y_;
+			uint32_t b_cy_max = static_cast<uint32_t>(cy_max) - chunks_y_;
 			glPushMatrix();
 			glTranslatef((float)size_x_, (float)size_y_, 0.0f);
 			renderSlice(tiles_tex, a_cx_min, a_cx_max, b_cy_min, b_cy_max, stateBlend);
@@ -165,7 +164,7 @@ void narf::client::World::render(narf::gl::Texture *tiles_tex, const narf::Camer
 		}
 
 		if (cy_min < 0) {
-			uint32_t h_cy_min = cy_min + chunks_y_;
+			uint32_t h_cy_min = static_cast<uint32_t>(cy_min) + chunks_y_;
 			uint32_t h_cy_max = chunks_y_;
 			glPushMatrix();
 			glTranslatef((float)size_x_, -(float)size_y_, 0.0f);
@@ -175,7 +174,7 @@ void narf::client::World::render(narf::gl::Texture *tiles_tex, const narf::Camer
 	}
 
 	if (cx_min < 0) {
-		uint32_t e_cx_min = cx_min + chunks_x_;
+		uint32_t e_cx_min = static_cast<uint32_t>(cx_min) + chunks_x_;
 		uint32_t e_cx_max = chunks_x_;
 		glPushMatrix();
 		glTranslatef(-(float)size_x_, 0.0f, 0.0f);
@@ -184,7 +183,7 @@ void narf::client::World::render(narf::gl::Texture *tiles_tex, const narf::Camer
 
 		if ((uint32_t)cy_max > chunks_y_) {
 			uint32_t d_cy_min = 0;
-			uint32_t d_cy_max = cy_max - chunks_y_;
+			uint32_t d_cy_max = static_cast<uint32_t>(cy_max) - chunks_y_;
 			glPushMatrix();
 			glTranslatef(-(float)size_x_, (float)size_y_, 0.0f);
 			renderSlice(tiles_tex, e_cx_min, e_cx_max, d_cy_min, d_cy_max, stateBlend);
@@ -192,7 +191,7 @@ void narf::client::World::render(narf::gl::Texture *tiles_tex, const narf::Camer
 		}
 
 		if (cy_min < 0) {
-			uint32_t f_cy_min = cy_min + chunks_y_;
+			uint32_t f_cy_min = static_cast<uint32_t>(cy_min) + chunks_y_;
 			uint32_t f_cy_max = chunks_y_;
 			glPushMatrix();
 			glTranslatef(-(float)size_x_, -(float)size_y_, 0.0f);
@@ -203,7 +202,7 @@ void narf::client::World::render(narf::gl::Texture *tiles_tex, const narf::Camer
 
 	if ((uint32_t)cy_max > chunks_y_) {
 		uint32_t c_cy_min = 0;
-		uint32_t c_cy_max = cy_max - chunks_y_;
+		uint32_t c_cy_max = static_cast<uint32_t>(cy_max) - chunks_y_;
 		glPushMatrix();
 		glTranslatef(0.0f, (float)size_y_, 0.0f);
 		renderSlice(tiles_tex, mid_cx_min, mid_cx_max, c_cy_min, c_cy_max, stateBlend);
@@ -211,7 +210,7 @@ void narf::client::World::render(narf::gl::Texture *tiles_tex, const narf::Camer
 	}
 
 	if (cy_min < 0) {
-		uint32_t g_cy_min = cy_min + chunks_y_;
+		uint32_t g_cy_min = static_cast<uint32_t>(cy_min) + chunks_y_;
 		uint32_t g_cy_max = chunks_y_;
 		glPushMatrix();
 		glTranslatef(0.0f, -(float)size_y_, 0.0f);
